refactor: simplify app structure
