--- src/main/factories/app/index.ts
@@ -1,2 +1,0 @@
-export * from './instance'
-export * from './setup'
