--- src/main/factories/index.ts
@@ -1,3 +1,0 @@
-export * from './ipcs/register-window-creation'
-export * from './windows/create'
-export * from './app'
