+++ src/main/index.ts
@@ -1,11 +1,10 @@
 import { app } from 'electron'
 
-import { makeAppSetup, makeAppWithSingleInstanceLock } from './factories'
-import { MainWindow, registerAboutWindowCreationByIPC } from './windows'
+import { makeAppWithSingleInstanceLock } from './factories/app/instance'
+import { makeAppSetup } from './factories/app/setup'
+import { MainWindow } from './windows/main'
 
 makeAppWithSingleInstanceLock(async () => {
   await app.whenReady()
   await makeAppSetup(MainWindow)
-
-  registerAboutWindowCreationByIPC()
 })
