--- src/main/windows/About/ipcs/index.ts
@@ -1,1 +1,0 @@
-export * from './register-window-creation'
