--- src/main/windows/index.ts
@@ -1,2 +1,0 @@
-export * from './About'
-export * from './Main'
