--- src/preload/ipcs/windows/about/create.ts
@@ -1,9 +1,0 @@
-import { ipcRenderer } from 'electron'
-
-import { IPC } from 'shared/constants'
-
-export function createAboutWindow() {
-  const channel = IPC.WINDOWS.ABOUT.CREATE_WINDOW
-
-  ipcRenderer.invoke(channel)
-}
