--- src/preload/ipcs/windows/index.ts
@@ -1,2 +1,0 @@
-export * from './about/when-close'
-export * from './about/create'
