--- src/renderer/components/Button/styles.module.sass
@@ -1,23 +1,0 @@
-.button
-  margin: 1rem 0
-  padding: 15px 35px
-  width: max-content
-  color: white
-  background-image: linear-gradient(to right, #DD5E89 0%, #F7BB97  51%, #DD5E89  100%)
-  background-size: 200% auto
-  border-color: #ffdacd
-  border: 2px groove
-  border-radius: 10px
-  box-shadow: 0 0 20px #eee
-  text-align: center
-  text-transform: uppercase
-  cursor: pointer
-  transition: 0.5s
-
-  &:hover
-    color: #fff
-    background-position: right center
-
-.button:nth-child(even)
-  background-image: linear-gradient(to right, #e3a7ff 0%, #8966f4 51%, #dd5eb7 100%)
-  border-color: #f2d5ff
