--- src/renderer/components/index.ts
@@ -1,3 +1,0 @@
-export * from './Container'
-export * from './Heading'
-export * from './Button'
