+++ src/renderer/index.tsx
@@ -1,7 +1,6 @@
 import ReactDom from 'react-dom/client'
 import React from 'react'
 
-import { WindowStoreProvider } from './store'
 import { AppRoutes } from './routes'
 
 import 'resources/styles/globals.sass'
@@ -9,8 +8,6 @@ import 'resources/styles/globals.css'
 
 ReactDom.createRoot(document.querySelector('app') as HTMLElement).render(
   <React.StrictMode>
-    <WindowStoreProvider>
-      <AppRoutes />
-    </WindowStoreProvider>
+    <AppRoutes />
   </React.StrictMode>
 )
