+++ src/renderer/routes.tsx
@@ -2,7 +2,7 @@ import { Route } from 'react-router-dom'
 
 import { Router } from 'shared/lib/electron-router-dom'
 
-import { MainScreen, AboutScreen, AnotherScreen } from 'renderer/screens'
+import { MainScreen } from './screens/Main'
 
 export function AppRoutes() {
   return (
@@ -10,10 +10,8 @@ export function AppRoutes() {
       main={
         <>
           <Route path="/" element={<MainScreen />} />
-          <Route path="/anotherScreen" element={<AnotherScreen />} />
         </>
       }
-      about={<Route path="/" element={<AboutScreen />} />}
     />
   )
 }
