--- src/renderer/screens/About/styles.module.sass
@@ -1,2 +1,0 @@
-.subtitle
-  width: 260px
