--- src/renderer/screens/index.ts
@@ -1,3 +1,0 @@
-export * from './Another'
-export * from './About'
-export * from './Main'
