--- src/renderer/store/index.tsx
@@ -1,40 +1,0 @@
-import { useContext, createContext, useState } from 'react'
-
-export interface WindowStore {
-  about: {
-    isOpen: boolean
-    setAboutWindowState: (value: boolean) => void
-  }
-}
-
-const WindowStoreContext = createContext({} as WindowStore)
-
-export function useWindowStore() {
-  return useContext(WindowStoreContext)
-}
-
-export function WindowStoreProvider({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  const [state, setState] = useState({
-    about: { isOpen: false, setAboutWindowState },
-  })
-
-  function setAboutWindowState(value: boolean) {
-    setState((state) => ({
-      ...state,
-      about: {
-        ...state.about,
-        isOpen: value,
-      },
-    }))
-  }
-
-  return (
-    <WindowStoreContext.Provider value={state}>
-      {children}
-    </WindowStoreContext.Provider>
-  )
-}
