--- src/shared/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './environment'
-export * from './platform'
-export * from './ipc'
