--- src/shared/constants/ipc.ts
@@ -1,8 +1,0 @@
-export const IPC = {
-  WINDOWS: {
-    ABOUT: {
-      CREATE_WINDOW: 'windows: create-about-window',
-      WHEN_WINDOW_CLOSE: 'windows: when-about-window-close',
-    },
-  },
-}
