--- src/shared/index.ts
@@ -1,3 +1,0 @@
-export * as constants from './constants'
-export * as utils from './utils'
-export * as types from './types'
